Fixed import error under heroku envs
