+++ open_discussions/utils_test.py
@@ -6,7 +6,6 @@ import pytest
 import pytz
 
 from open_discussions.utils import (
-    assert_not_raises,
     now_in_utc,
     is_near_now,
     normalize_to_start_of_day,
@@ -76,27 +75,3 @@ def test_chunks_iterable():
     for chunk in chunk_output:
         range_list += chunk
     assert range_list == list(range(count))
-
-
-def test_assert_not_raises_none():
-    """
-    assert_not_raises should do nothing if no exception is raised
-    """
-    with assert_not_raises():
-        pass
-
-
-def test_assert_not_raises_exception(mocker):
-    """assert_not_raises should fail the test"""
-    # Here there be dragons
-    fail_mock = mocker.patch('pytest.fail', autospec=True)
-    with assert_not_raises():
-        raise TabError()
-    assert fail_mock.called is True
-
-
-def test_assert_not_raises_failure():
-    """assert_not_raises should reraise an AssertionError"""
-    with pytest.raises(AssertionError):
-        with assert_not_raises():
-            assert 1 == 2
